Updated tentacle script
added tentacles, imported wing material for tentacles, added Movement script (incomplete)
