+++ Assets/Materials/Wing 1.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: be21071c75fa729499ec1e359563fa25
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
