+++ Assets/Materials/Wing 2.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 966765d83d6fb6a4287a0f166611ad93
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
