+++ Assets/Scripts/Entities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a70b5b8b46532d141bfa40e680b97be2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
