+++ Assets/Scripts/Entities/RotateToTarget.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2911f866f610d814ba1e780c9b2c61bb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
