+++ Assets/Scripts/Entities/Tentacle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dbd029d3a315afe47995624b9288d802
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
