+++ Assets/Sprites/Entities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4e651a545c2944041b805d0473409e0a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
